Fragments Merged As Original View
