+++ app/src/main/java/com/lerex/tr/ViewManager.java
@@ -35,9 +35,7 @@ public class ViewManager extends AppCompatActivity {
 
 
 
-        tinydb = new TinyDB(this);
-        tinydb.deleteImage("Movies");
-        new GetmovieResults().execute();
+
         Button sell=findViewById(R.id.btnSell);
         Button buy=findViewById(R.id.btnBuy);
         Button bar=findViewById(R.id.btnVM);
@@ -76,77 +74,5 @@ public class ViewManager extends AppCompatActivity {
     }
 
 
-    private class GetmovieResults extends AsyncTask<Void, Void, Void> {
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            Toast.makeText(ViewManager.this, "Json Data is downloading", Toast.LENGTH_LONG).show();
 
-        }
-
-        @Override
-        protected Void doInBackground(Void... arg0) {
-            HttpHandler sh = new HttpHandler();
-            // Making a request to url and getting response
-            String url = "https://api.themoviedb.org/3/movie/now_playing?region=IN&page=1&language=en-US&api_key=cdb6543f56d4ae849f71ed220c46a080";
-            String jsonStr = sh.makeServiceCall(url);
-
-            Log.e(TAG, "Response from url: " + jsonStr);
-            if (jsonStr != null) {
-                try {
-                    JSONObject jsonObj = new JSONObject(jsonStr);
-
-                    // Getting JSON Array node
-                    JSONArray movieResults = jsonObj.getJSONArray("results");
-
-                    // looping through All movieResults
-                    for (int i = 0; i < movieResults.length(); i++) {
-                        JSONObject c = movieResults.getJSONObject(i);
-                        String title = c.getString("title");
-
-                        // tmp hash map for single contact
-                        HashMap<String, String> moviesReleased = new HashMap<>();
-
-                        // adding each child node to HashMap key => value
-                        moviesReleased.put("title", title);
-                        movies.add(moviesReleased.get("title"));
-
-                    }
-                    tinydb.putListString("Movies",movies);
-                    System.out.println(tinydb.getListString("Movies"));
-                } catch (final JSONException e) {
-                    Log.e(TAG, "Json parsing error: " + e.getMessage());
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            Toast.makeText(getApplicationContext(), "Json parsing error: " + e.getMessage(),
-                                    Toast.LENGTH_LONG).show();
-                        }
-                    });
-
-                }
-
-            } else {
-                Log.e(TAG, "Couldn't get json from server.");
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        Toast.makeText(getApplicationContext(),
-                                "Couldn't get json from server. Check LogCat for possible errors!",
-                                Toast.LENGTH_LONG).show();
-                    }
-                });
-            }
-
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Void result) {
-            super.onPostExecute(result);
-            //To Add Modifications For Drop Down List Box
-            Toast.makeText(ViewManager.this, "Json Data downloaded", Toast.LENGTH_LONG).show();
-        }
-
-    }
 }
