+++ app/src/main/java/com/lerex/tr/buyerActivity.java
@@ -1,22 +1,29 @@
 package com.lerex.tr;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.ArrayAdapter;
 import android.widget.AutoCompleteTextView;
 import android.widget.Button;
-import android.widget.ListView;
+import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -26,47 +33,35 @@ import com.google.firebase.database.ValueEventListener;
 import java.util.ArrayList;
 import java.util.Objects;
 
+public class buyerActivity extends Fragment{
 
-public class buyerActivity extends AppCompatActivity {
-
-
-
-    private String TAG = buyerActivity.class.getSimpleName();
     protected AutoCompleteTextView search;
-    private RecyclerView.Adapter adapter;
     private ArrayList<TicketDetails> buylist;
+    private RecyclerView.Adapter adapter;
     private RecyclerView.LayoutManager layoutManager;
     private RecyclerView recyclerView;
+    private TextView tv;
+    private Button searchbtn;
 
 
-
+    @Nullable
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
 
-        requestWindowFeature(Window.FEATURE_NO_TITLE);
-        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-                WindowManager.LayoutParams.FLAG_FULLSCREEN);//FullScreening The Application
-
-        setContentView(R.layout.activity_buyer);
-
-
-
-        TinyDB tinydb = new TinyDB(this);//Shared Preference To Get Localized Data
+        View view = inflater.inflate(R.layout.activity_buyer, null);
+        TinyDB tinydb = new TinyDB(getActivity());//Shared Preference To Get Localized Data
         ArrayList<String> tickets = tinydb.getListString("Movies");
 
-        search=findViewById(R.id.tvSearch);
-        Button search1=findViewById(R.id.btnSearch);
+        search=view.findViewById(R.id.tvSearch);
+        searchbtn=view.findViewById(R.id.btnSearch);
 
-
-        ArrayAdapter<String> searchAdapter = new ArrayAdapter<>(buyerActivity.this, android.R.layout.simple_dropdown_item_1line, tickets);
+        ArrayAdapter<String> searchAdapter = new ArrayAdapter<>(Objects.requireNonNull(getActivity()), android.R.layout.simple_dropdown_item_1line, tickets);
         search.setAdapter(searchAdapter);
-        recyclerView = findViewById(R.id.rvAvailable);
+        recyclerView = view.findViewById(R.id.rvAvailable);
         recyclerView.setHasFixedSize(true);
 
-        layoutManager = new LinearLayoutManager(this);
+        layoutManager = new LinearLayoutManager(getActivity());
         recyclerView.setLayoutManager(layoutManager);
-
 
         buylist=new ArrayList<>();
         adapter = new BuyerListView(buylist);
@@ -76,22 +71,15 @@ public class buyerActivity extends AppCompatActivity {
         ItemTouchHelper itemTouchhelper = new ItemTouchHelper(swipeController);
         itemTouchhelper.attachToRecyclerView(recyclerView);
 
-        search1.setOnClickListener(new View.OnClickListener() {
+        searchbtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 GetAvailable();
-
             }
         });
 
-
+        return view;
     }
-
-    /*@Override
-    public void onBackPressed() {
-        startActivity(new Intent(this,ViewManager.class));
-    }*/
-
 
     protected void GetAvailable(){
 
@@ -130,16 +118,21 @@ public class buyerActivity extends AppCompatActivity {
             }
         });
 
+        //Is this Necessary And If So Why is It
+
+        /*assert getFragmentManager() != null;
+        FragmentTransaction ft = getFragmentManager().beginTransaction();
+        ft.detach(buyerActivity.this).attach(buyerActivity.this).commit();*/
     }
-    ItemTouchHelper.SimpleCallback itemTouchHelperCallback=new ItemTouchHelper.SimpleCallback(0,ItemTouchHelper.RIGHT|ItemTouchHelper.LEFT) {
-        @Override
-        public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
-            return false;
-        }
 
-        @Override
-        public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-
+    @Override
+    public void onResume() {
+        super.onResume();
+        View viewNavigation= Objects.requireNonNull(getActivity()).findViewById(R.id.bottomNavigationView);
+        if(viewNavigation instanceof BottomNavigationView){
+            BottomNavigationView bottomNavView=(BottomNavigationView)viewNavigation;
+            bottomNavView.setSelectedItemId(R.id.buyer);
         }
-    };
+    }
 }
+
