+++ app/src/main/res/layout/activity_buyer.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#7A88E6FF"
+    android:background="@drawable/bg3"
     tools:context=".buyerActivity">
 
     <AutoCompleteTextView
@@ -19,7 +19,7 @@
         android:textCursorDrawable="@android:color/background_dark"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.229"
+        app:layout_constraintHorizontal_bias="0.186"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.102"
@@ -45,7 +45,9 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tvSearch" />
+        app:layout_constraintTop_toBottomOf="@+id/tvSearch"
+        app:layout_constraintVertical_bias="0.15"
+        />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
