--- app/src/main/res/layout/activity_buyer1.xml
@@ -1,65 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout
-    android:id="@+id/buyerLa"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/bg3"
-    tools:context=".sellActivity">
-
-    <AutoCompleteTextView
-        android:id="@+id/tvSearch1"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:completionThreshold="1"
-        android:hint="Select a movie"
-        android:textColor="#000000"
-        android:textColorHighlight="#19232D"
-        android:textColorLink="#19232D"
-        android:textCursorDrawable="@android:color/background_dark"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.186"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.102"
-        tools:ignore="HardcodedText" />
-
-    <Button
-        android:id="@+id/btnSearch1"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:background="@drawable/search"
-        android:backgroundTint="#19232D"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.63"
-        app:layout_constraintStart_toEndOf="@+id/tvSearch1"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.101" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvAvailable1"
-        android:layout_width="350dp"
-        android:layout_height="498dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.491"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tvSearch1"
-        app:layout_constraintVertical_bias="0.14" />
-
-    <TextView
-        android:id="@+id/tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toTopOf="@+id/rvAvailable1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.25" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
