+++ app/src/main/res/layout/activity_sell.xml
@@ -8,23 +8,6 @@
     android:background="@drawable/bg3"
     tools:context=".sellActivity">
 
-    <!--<ListView
-        android:id="@+id/lvTickets"
-        android:layout_width="350dp"
-        android:layout_height="524sp"
-        android:layout_marginTop="25dp"
-        android:divider="@android:color/transparent"
-        android:dividerHeight="10.0sp"
-        app:layout_constraintBottom_toTopOf="@+id/btnAdd"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.495"
-        android:focusableInTouchMode="false"
-        android:isScrollContainer="false"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0" />-->
-
-
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
         android:layout_width="match_parent"
@@ -43,9 +26,9 @@
         android:textColor="#FFFFFF"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.953" />
+        app:layout_constraintVertical_bias="0.885" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
